+++ src/stores/createStackStore/index.ts
@@ -1,30 +1,16 @@
 import {
   createStore,
-  YobtaIdleEvent,
-  YobtaReadyEvent,
+  YobtaStore,
   YobtaStorePlugin,
-  YobtaTransitionEvent,
 } from '../createStore/index.js'
 
-export type YobtaStackStore<Item, Overloads extends any[] = any[]> = {
-  add(member: Item, ...overloads: any[]): boolean
-  last(): Item
-  observe(
-    observer: (state: ReadonlySet<Item>, ...overloads: Overloads) => void,
-  ): VoidFunction
-  on(
-    topic: YobtaReadyEvent | YobtaIdleEvent | YobtaTransitionEvent,
-    subscriber: (state: ReadonlySet<Item>) => void,
-  ): VoidFunction
-  remove(member: Item, ...overloads: any[]): boolean
+export type YobtaStackStore<Item, Overloads extends any[] = any[]> = Omit<
+  YobtaStore<Item | undefined, Overloads>,
+  'next'
+> & {
+  push(member: Item, ...overloads: any[]): number
+  pop(...overloads: any[]): Item | undefined
   size(): number
-}
-
-interface YobtaStackStoreFactory {
-  <Item, Overloads extends any[] = any[]>(
-    initialState: Set<Item> | Item[],
-    ...plugins: YobtaStorePlugin<ReadonlySet<Item>, Overloads>[]
-  ): YobtaStackStore<Item, Overloads>
 }
 
 /**
@@ -33,39 +19,31 @@ interface YobtaStackStoreFactory {
  * const store = createStackStore([])
  * @documentation {@link https://github.com/yobta/stores/tree/master/src/stores/createStackStore/index.md}
  */
-export const createStackStore: YobtaStackStoreFactory = <
-  Item,
-  Overloads extends any[],
->(
-  initialState?: Set<Item> | Item[],
-  ...plugins: YobtaStorePlugin<ReadonlySet<Item>, Overloads>[]
-) => {
-  let { last, observe, next, on } = createStore<ReadonlySet<Item>, Overloads>(
-    new Set(initialState),
-    ...plugins,
-  )
+export const createStackStore = <Item, Overloads extends any[]>(
+  initialState: Item[] = [],
+  ...plugins: YobtaStorePlugin<Item | undefined, Overloads>[]
+): YobtaStackStore<Item | undefined, Overloads> => {
+  let stack = [...initialState]
+
+  let value = (): Item | undefined => stack[stack.length - 1]
+
+  let { next, ...store } = createStore(value(), ...plugins)
+
   return {
-    add(item: Item, ...overloads: Overloads) {
-      let state = last()
-      if (!state.has(item)) {
-        let nextState = new Set([item, ...state])
-        next(nextState, ...overloads)
-        return true
+    ...store,
+    push(item: Item, ...overloads: Overloads) {
+      let size = stack.push(item)
+      next(value(), ...overloads)
+      return size
+    },
+    pop(...overloads: Overloads) {
+      let lastLength = stack.length
+      let last = stack.pop()
+      if (lastLength !== stack.length) {
+        next(value(), ...overloads)
       }
-      return false
+      return last
     },
-    last() {
-      let [first] = last()
-      return first
-    },
-    observe,
-    on,
-    remove(item: Item, ...overloads: Overloads) {
-      let state = new Set(last())
-      let result = state.delete(item)
-      if (result) next(state, ...overloads)
-      return result
-    },
-    size: () => last().size,
+    size: () => stack.length,
   }
 }